requirements.txt
altgraph==0.17.3
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
icmplib==3.0.3
idna==3.4
isodate==0.6.1
lxml==4.9.2
mccabe==0.7.0
onvif-zeep==0.2.12
pexpect==4.8.0
platformdirs==2.6.0
ptyprocess==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
RouterOS-api==0.17.1.dev0
six==1.16.0
tqdm==4.64.1
urllib3==1.26.13
zeep==4.2.1
